Move /src/bin to /benches
