--- src/bin/extra_impls/mod.rs
@@ -1,1 +1,0 @@
-pub mod mpsc_queue;
